+++ akiba/akiba-wsclient/src/main/java/pl/akiba/wsclient/client/DefaultClient.java
@@ -39,7 +39,7 @@ public class DefaultClient {
         this.mapper = new ObjectMapper();
     }
 
-    protected StringBuilder prepareBasicUrl(int userId) {
+    protected StringBuilder prepareBasicUrl(long userId) {
         return new StringBuilder(address).append("/").append(userId);
     }
